requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
pydub>=0.25.1
pillow>=10.0.0
moviepy>=1.0.3
openai>=1.3.0
tqdm>=4.66.1
streamlit>=1.31.0
gtts>=2.5.0
zhipuai>=2.0.0
pymongo>=4.6.1 